+++ CHANGES.md
@@ -1,5 +1,8 @@
 ## Changelog ##
 
+### 3.2-BETA1 ###
+* Added: Filter hook to disable Enlighter on selected pages - feature requested on [WordPress.org Forums](https://wordpress.org/support/topic/best-way-to-dequeue-enlighter-plugin?replies=2) #43
+
 ### 3.1 ###
 * Added: [EnlighterJS v2.10.1](http://enlighterjs.org/)
 * Added: About/News Page which is shown on plugin activation/upgrade